feat: Context API 공통 훅으로 변경
- 공통 훅 내에서 initialState 세팅
- test.tsx에서 주입받아서도 활용가능하도록 수정
