+++ src/refactoring/App.tsx
@@ -1,76 +1,29 @@
 import { AdminPage } from "@/refactoring/components/AdminPage";
 import { CartPage } from "@/refactoring/components/CartPage";
-import { useCoupons, useProducts, useService } from "@/refactoring/hooks";
-import { Coupon, PERMISSION_TYPE, Product } from "@/types";
-
-const initialProducts: Product[] = [
-  {
-    id: "p1",
-    name: "상품1",
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-  },
-  {
-    id: "p2",
-    name: "상품2",
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-  },
-  {
-    id: "p3",
-    name: "상품3",
-    price: 30000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.2 }],
-  },
-];
-
-const initialCoupons: Coupon[] = [
-  {
-    name: "5000원 할인 쿠폰",
-    code: "AMOUNT5000",
-    discountType: "amount",
-    discountValue: 5000,
-  },
-  {
-    name: "10% 할인 쿠폰",
-    code: "PERCENT10",
-    discountType: "percentage",
-    discountValue: 10,
-  },
-];
+import { useService } from "@/refactoring/hooks";
+import { PERMISSION_TYPE } from "@/types";
+import { Providers } from "./provider";
 
 // 서버에서 사용자에게 관리자 권한이 있는 지를 더미 테스트
 const SERVER_USER_PERMISSION = PERMISSION_TYPE.ADMIN;
 
 const App = () => {
-  const { products, updateProduct, addProduct } = useProducts(initialProducts);
-  const { coupons, addCoupon } = useCoupons(initialCoupons);
   const { isAdminView, toggleService } = useService(SERVER_USER_PERMISSION);
 
   return (
-    <div className="min-h-screen bg-gray-100">
-      <nav className="bg-blue-600 text-white p-4">
-        <div className="container mx-auto flex justify-between items-center">
-          <h1 className="text-2xl font-bold">쇼핑몰 관리 시스템</h1>
-          <button onClick={toggleService} className="bg-white text-blue-600 px-4 py-2 rounded hover:bg-blue-100">
-            {isAdminView ? "장바구니 페이지로" : "관리자 페이지로"}
-          </button>
-        </div>
-      </nav>
-      <main className="container mx-auto mt-6">
-        {isAdminView ? (
-          <AdminPage products={products} coupons={coupons} onProductUpdate={updateProduct} onProductAdd={addProduct} onCouponAdd={addCoupon} />
-        ) : (
-          <CartPage products={products} coupons={coupons} />
-        )}
-      </main>
-    </div>
+    <Providers>
+      <div className="min-h-screen bg-gray-100">
+        <nav className="bg-blue-600 text-white p-4">
+          <div className="container mx-auto flex justify-between items-center">
+            <h1 className="text-2xl font-bold">쇼핑몰 관리 시스템</h1>
+            <button onClick={toggleService} className="bg-white text-blue-600 px-4 py-2 rounded hover:bg-blue-100">
+              {isAdminView ? "장바구니 페이지로" : "관리자 페이지로"}
+            </button>
+          </div>
+        </nav>
+        <main className="container mx-auto mt-6">{isAdminView ? <AdminPage /> : <CartPage />}</main>
+      </div>
+    </Providers>
   );
 };
